third_party/WebKit/LayoutTests/fast/dom/navigatorcontentutils/is-protocol-handler-registered-expected.txt

This test makes sure that navigator.isProtocolHandlerRegistered throws the proper exceptions and returns the default state of handler.

PASS window.navigator.isProtocolHandlerRegistered is defined.
PASS Invalid url "%S" threw SyntaxError exception.
PASS Invalid scheme "mailto:" falied.
PASS Invalid scheme "ssh:/" falied.
PASS Invalid scheme "magnet:+" falied.
PASS Invalid scheme "tel:sip" falied.
PASS window.navigator.isProtocolHandlerRegistered returns "new" state
PASS successfullyParsed is true

TEST COMPLETE
